move analytics to separate dir
