+++ src/lib/abacus/logger.ts
@@ -3,7 +3,10 @@ import type { AbacusLoggingProtocol } from '@/constants/abacus';
 
 class AbacusLogger implements Omit<AbacusLoggingProtocol, '__doNotUseOrImplementIt'> {
   d(tag: string, message: string) {
-    if (import.meta.env.VITE_ENABLE_ABACUS_LOGGING) {
+    if (
+      import.meta.env.VITE_ENABLE_ABACUS_LOGGING ||
+      import.meta.env.VITE_ABACUS_LOG_LEVEL === 'debug'
+    ) {
       console.log(`${tag}: ${message}`);
     }
   }
