+++ src/lib/analytics/analytics.ts
@@ -6,6 +6,8 @@ import {
   type AnalyticsUserProperty,
 } from '@/constants/analytics';
 
+import { dd } from './datadog';
+
 const DEBUG_ANALYTICS = false;
 
 export const identify = (property: AnalyticsUserProperty) => {
@@ -19,6 +21,7 @@ export const identify = (property: AnalyticsUserProperty) => {
   const customEvent = customIdentifyEvent({
     detail: { property: propertyTypeToLog, propertyValue: property.payload },
   });
+  dd.setContextProperty(propertyTypeToLog, property.payload);
 
   globalThis.dispatchEvent(customEvent);
 };
